visualizer, minor: move files around
